Prefs: Separate webkitgtk 'server' and context handler from the window
